feat(ProtocolVersion): HTTP1인 프로토콜을 default로 생성하는 생성자 생성
- Protocol 과 Version을 잘 조합하여 가져오는지 테스트 작성
