+++ tomcat/src/test/java/org/apache/coyote/protocolVersion/ProtocolVersionTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.domain.protocolVersion;
+package org.apache.coyote.protocolVersion;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -10,8 +10,7 @@ class ProtocolVersionTest {
     @Test
     @DisplayName("Protocol 과 Version을 조합하여 가져올 수 있다.")
     void getCombinedProtocolVersion() {
-        String inputProtocolVersion = "HTTP/1.1";
-        ProtocolVersion protocolVersion = new ProtocolVersion(inputProtocolVersion);
+        ProtocolVersion protocolVersion = ProtocolVersion.ofHTTP1();
 
         assertEquals(protocolVersion.getCombinedProtocolVersion(), "HTTP/1.1");
     }
